+++ packages/react/src/components/Html.tsx
@@ -1,4 +1,4 @@
-import { HtmlExporters, HtmlDefaults } from "@unlayer-dev/exporters";
+import { HtmlExporters, HtmlDefaults } from "@unlayer/exporters";
 import type { HtmlValues } from "../types";
 import { createItemComponent, type ItemComponentProps } from "../utils/create-component";
 import { mapSemanticProps, type SemanticProps } from "../utils/semantic-props";