+++ packages/react/src/components/Social.tsx
@@ -1,4 +1,4 @@
-import { SocialExporters, SocialDefaults } from "@unlayer-dev/exporters";
+import { SocialExporters, SocialDefaults } from "@unlayer/exporters";
 import type { SocialValues, SocialIcon } from "../types";
 import { createItemComponent, type ItemComponentProps } from "../utils/create-component";
 import { mapSemanticProps, type SemanticProps } from "../utils/semantic-props";