+++ packages/react/src/components/Table.tsx
@@ -1,4 +1,4 @@
-import { TableExporters, TableDefaults } from "@unlayer-dev/exporters";
+import { TableExporters, TableDefaults } from "@unlayer/exporters";
 import type { TableValues } from "../types";
 import { createItemComponent, type ItemComponentProps } from "../utils/create-component";
 import { mapSemanticProps, type SemanticProps } from "../utils/semantic-props";