+++ packages/shared/tsup.config.ts
@@ -15,6 +15,6 @@ export default defineConfig({
     '@lexical/link',
     '@lexical/list',
     '@lexical/rich-text',
-    '@unlayer-dev/exporters',
+    '@unlayer/exporters',
   ],
 });